Update script.js
made a var to go on the next question when the right answer is click(let currentQuestionNum =0;)

and

added the click() for when you click the correct answer something. it is showed it is being click in the console but it is doing nothing on the screen at the moment.